+++ src/main/java/racing/RacingArena.java
@@ -1,0 +1,9 @@
+package racing;
+
+import racing.view.ResultView;
+
+public class RacingArena {
+    public static void main(String[] args) {
+        new ResultView().play(new CarRacing().start());
+    }
+}
